requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
PyYAML==6.0.2
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.0
requests==2.32.3
tqdm==4.66.4
nltk==3.9.1
beautifulsoup4==4.12.3
google-api-python-client==2.139.0
trafilatura==1.9.0
streamlit==1.37.0